Using sweetalert2

+++ frontend/src/components/book.tsx
@@ -3,6 +3,7 @@ import { useMutation } from "@apollo/client";
 import { DELETE_BOOK } from "../hooks/deleteBook";
 import LoadingBox from "./LoadingBox";
 import ErrorBox from "./Error";
+import Swal from "sweetalert2";
 
 interface BookProps {
   _id: Key;
@@ -20,7 +21,24 @@ const Book = ({ _id, title, author }: BookProps) => {
       <div className="buttons">
         <button className="update">update</button>
         <button className="delete" onClick={() => {
-          deleteBook({variables: {id: _id}})
+          Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+          }).then((result) => {
+            if (result.isConfirmed) {
+              deleteBook({variables: {id: _id}});
+              Swal.fire(
+                'Deleted!',
+                'Your file has been deleted.',
+                'success'
+              )
+            }
+          });
         }}>delete</button>
       </div>
       <span className="author"><small>by: {author}</small></span>
